+++ groups/bsl/bslma/bslma_autodeallocator.t.cpp
@@ -379,7 +379,7 @@ int TestAllocator::numOutstandingAllocations() const
         char **tmpSrc = srcArray.d_array_p;
         if (this == &srcArray) {
             // self-alias
-            int size = srcLength * sizeof(char *);
+            size_t size = srcLength * sizeof(char *);
             tmpSrc = (char **) d_allocator_p->allocate(size);
             memcpy(tmpSrc, d_array_p, size);
         }
@@ -455,7 +455,7 @@ int TestAllocator::numOutstandingAllocations() const
         char **tmpSrc = srcArray.d_array_p;
         if (this == &srcArray) {
             // self-alias
-            int size = srcLength * sizeof(char *);
+            size_t size = srcLength * sizeof(char *);
             tmpSrc = (char **) d_allocator_p->allocate(size);
             memcpy(tmpSrc, d_array_p, size);
         }
